De-virtualize region handle
All region handles are now bitmap backed.
